optimization of force loop
when accumulating the forces for each atom i into a set of
temporary variables before adding them to the main array,
we allow the compiler to keep data in registers longer.
as a positive side effect, we slightly increase our accuracy.
this should work for all simple, lennard-jones type potentials.