make fft header honor FFTW_SIZE define for fftw2
in fftw-2 the single and double precision API are the
same. to install both at the same time, one can select
to have the header and library files be installed with
an s- or d-letter prefix, so that both can be installed
into the same path. this define takes care of those cases.