more single precision FFT code
do more calculations in single precision, particularly
the accumulation of the charge density and the calculation
of the long range potential contribution to the forces.
this way we avoid converting back and forth between single
and double precision for several times.

port of the single prec changes also to pppm/cg and pppm/tip4p.