fix another apparent bug in compute atom/molecule
the code was not using the right way to determine whether
per atom data is available and would use the wrong stride.
also, some compute code seems dependant on data vectors/arrays
being initialized in compute classes, however they were not
properly set to NULL in the base class constructor for this
to work reliably.