avoid accessing already deleted LAMMPS component instances in destructor
when saving the colvars state from the destructor of the proxy module
some LAMMPS component instances have already been deleted, e.g. update.
thus we valgrind complains when accessing parameters/constants that
they provide (timestep, unit conversion factors) and we have to maintain
a cache of those values inside the proxy module. most are uncritical and
don't change, but the timestep would have to be checked at every setup.
TODO: we need to check, for variable time step simulations and refuse to
work in those cases.