Added -thermo cli option to dump thermo data to a file
For regression testing it is often useful to directly dump thermo data
to a file instead of trying to parse the log file or manually adding
fix print commands. The thermo option allows to specify a filename
where thermo lines are dumped to.

This also adds a thermo_log command to change the thermo file from the
input script.