Wrong length of the box in the z-direction
I am trying to convert a *car file to LAMMPS data file. The fifth line of the car file reads:
PBC   29.4000   32.4000  172.0000   48.5000   77.0000   63.5000 (P1)
The box dimensions that I get using the tool is 
   -14.700000000    14.700000000 xlo xhi
   -14.497936658    14.497936658 ylo yhi
   -83.634474945    83.634474945 zlo zhi
    14.456809145    38.691581347   108.059935409 xy xz yz
Using the relationships in section 6.12 it is expected that the length of the simulation box in the z-direction would be 128.1015687. Moreover, the initial dimension in the car file (a, b, c) cannot be recovered from these values. When I apply the patch I can recover the initial values.