Update MakeLists.c
modifications to account for the case where the files are converted to a LAMMPS force field without any bonded contributions like the reax ff. The variables  total_no_bonds, total_no_angles, total_no_dihedrals, total_no_oops, total_no_angle_angles are set to zero if the type of the force field corresponds to FF_TYPE_REAXFF. The rest of the code for the determination of bonds, angles, dihedrals & impropers are thus not executed.