Fix wrong size argument when checking for eos/table/rx string
This PR fixes the size argument passed to `strncmp(modify->fix[i]->style, "eos/table/rx", 3)` call as the `"eos/table/rx"` string literal has a length of 12.

I am not aware of this codebase and so this might be an incorrect fix. If this check is correct, I would suggest updating the `"eos/table/rx"` string to just `"eos"`.