ignore processors command when nprocs == 1
With this patch, LAMMPS running on one MPI process will ignore a "processors" command, warning that it has done so.

Prior behavior causes LAMMPS to crash with (e.g.) "processors * * 2" while running on one process. This forces "processors" commands to be commented out for small-scale one-process testing (such as in unit tests or for debugging), even though LAMMPS can run unambiguously on one process. Forgetting to then uncomment a "processors" command for production can result in significant time lost.

This patch would allow production scripts to be tested on a single processor and seamlessly run in production.