Bugfix fep tools
If the free energy is divided by the volume, this should be done
outside the logarithm. 'tok[2]' is an average over exp(-k_b T Delta U),
i.e., division with the volume outside the logarithm is equivalent to
averaging over exp(-k_b T Delta U / volume).
In fdti.py the sum is divided by 'i' instead of 'i-1' because i=0 after
the first contribution to the sum.