Pair style with Gaussian width
This MR moves all pair interactions of the fix electrode to the new pair style lj/cut/coul/long/gauss.
Matrix  and vector computations are done in methods of the new virtual class ElectrodePair, in analogy to ElectrodeKSpace. Using this functionality is optional and is activated by the new "pair" keyword of the electrode fix. There is an example input "in.pair" in the graph-il example.

Separation of the pair interactions from the electrode fix will allow for more pair styles in the package, e.g., TIP4p and damped Coulomb interactions w/o long-range.