lj/cut/coul/wolf/gauss pair style
A new pair style, lj/cut/coul/wolf/gauss, with truncated Coulomb interactions is added to the ELECTRODE package. The pair style implements the ElectrodPair virtual class, i.e., it can be used with the pair command of `fix electrode`. Using a kspace style with 'fix electrode' is now optional. If the new pair style is used, the Wolf summation is performed with a correction for Gaussian charges.