Refactor electrode package
The electrode package is refactored, to make the code of the electrode fixes more readable.
A new parent class `ChargeSolver` is created to handle the solution of the constant potential method, i.e. minimization of the Coulomb energy as function of electrode charges.
ElectrodeInv and ElectrodeCG are two implementations of the solver class, for the matrix inversion and conjugate gradient method, respectively.
Management of global taglists is moved to the new `ElectrodeTaglist` class.
The `symm` keyword is deprecated and replaced by the `qtotal` keyword.