Predictor for conjugate gradient in fix electrode
A predictor is added to the conjugate gradient (CG) algorithm. Atom charges of previous time steps are stored in an atom/property array. These prior charges are then used to extrapolate the current charges as a starting point for the CG. The average number of CG steps is decreased by up to 60% (quartz example). If the predictor is used, an atom/property fix is created by the ElectrodCG class.

In order to set the number of prior charges, the keyword *predictor* is used. The default is `predictor 1`. To turn the new feature off, use `predictor 0`.